requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
